check_merged: only check 10 commits for merge
This had pathological behavior if the tip of a branch had a subject that
would match a large number of commits. My working example had a branch
whose commit subject was just `f`, and `git log --grep f` turns up an
awful lot of commits, which means we'd run diff-tree on an enormous
number of commits.

Fix this by limiting our log grep to 10 commits. In the normal case of a
sensible commit subject, this should easily find it. In the problematic
case, where some monster has written a single-letter commit message, we
won't spend forever looking for the merge; instead we'll just warn and
assume it's not merged.